backend/requirements.txt
anthropic==0.40.0
colpali_engine==0.3.4
fastapi==0.115.5
numpy==2.1.3
Pillow==11.0.0
pydantic==2.10.2
python-dotenv==1.0.1
pyvespa==0.50.0
supabase==2.10.0
torch==2.5.1
